rust (clippy)

+++ farmer-school-bevy/src/main.rs
@@ -139,7 +139,7 @@ fn screen_to_world(
     camera_transform: &GlobalTransform,
     screen_position: Vec2,
 ) -> Option<Vec2> {
-    let window_size = Vec2::new(window.width() as f32, window.height() as f32);
+    let window_size = Vec2::new(window.width(), window.height());
     let ndc = (screen_position / window_size) * 2.0 - Vec2::ONE;
     let ndc_to_world = camera_transform.compute_matrix() * camera.projection_matrix().inverse();
     let world_position = ndc_to_world.project_point3(ndc.extend(-1.0));
